Create bruxbot
One of the current issues with brux-gdk is that it's built with a normal distro that has a somewhat new toolchain.

That's usually perfectly fine, but it quickly breaks when you want to support distros with older toolchains.

bruxbot solves that problem by building it's own toolchain with specific software versions for compatibility.

Keep in mind that this is still a work-in-progress though: bruxbot currently doesn't work properly.

That should be fixed in future commits though.

Signed-off-by: hexaheximal <hexaheximal@proton.me>